requirements.txt
numpy==1.10.4
Pillow==3.1.1
pyOSC==0.3.5b5294
pyserial==3.0.1
scipy==0.17.0
